+++ CHANGES.rst
@@ -6,6 +6,7 @@ nmea-kindle-panel changelog
 in progress
 ===========
 - UDP receiver based on ``asyncio-dgram`` and ``pynmea2``
+- Add data model classes and core display rendering routines
 
 
 2022-07-xx 0.0.0